renaming skaffold template files and updating skaffold template generator
Summary:
Update naming scheme for skaffold template files from .tmpl to .skfld.tmpl
in light of reuse of .tmpl extension

Test Plan:
pull,
     ```
    rm -rf skaffold_build/
    make skaffold-dev
     ```

Reviewers: michelle, zasgar, kgandhi

Reviewed By: michelle

Differential Revision: https://phab.corp.pixielabs.ai/D112

GitOrigin-RevId: 17e7e6dce44e78777bd97215ef56d0024fd1535f

+++ utils/clutils.go
@@ -64,8 +64,15 @@ func RunCmd(cmd *exec.Cmd) error {
 		return err
 	}
 
+	counter := 0
 	addSignalInterruptCatch(func() {
-		cmd.Process.Kill()
+		// special kill switch in case keyboard interrupt is hit 3 times.
+		// otherwise, allow for graceful cleanup of command
+		// via keyboard interrupt
+		if counter > 3 {
+			cmd.Process.Kill()
+		}
+		counter++
 	})
 
 	err = cmd.Wait()
